+++ include/material/Material.hpp
@@ -14,22 +14,22 @@ class Material {
     public:
         Material() = default;
        ~Material();
-        Material(Color c, double k_amb, double k_dif, double k_spec,  double exp, double k_trans = 0, double index_ref = 0/*, IShade* luminance = nullptr*/);
+        Material(Color c, float k_amb, float k_dif, float k_spec,  float exp, float k_trans = 0, float index_ref = 0/*, IShade* luminance = nullptr*/);
 
-        Material&      turn_into_matte(double k_amb, double k_dif);
-        Material&    turn_into_plastic(double k_spec, double exp);
-        Material& turn_into_reflective(double k_trans, double index_ref);
+        Material&      turn_into_matte(float k_amb, float k_dif);
+        Material&    turn_into_plastic(float k_spec, float exp);
+        Material& turn_into_reflective(float k_trans, float index_ref);
         Material&         change_color(Color c);
         
 
 
     private:
         Color   color;
-        double  k_amb;
-        double  k_dif;
-        double  k_spec;
-        double  k_trans;
-        double  exp;
-        double  index_ref;
+        float  k_amb;
+        float  k_dif;
+        float  k_spec;
+        float  k_trans;
+        float  exp;
+        float  index_ref;
 };
