+++ src/math/Vector4.cpp
@@ -1,14 +1,14 @@
 #include "../../include/math/Vector4.hpp"
 #include <cmath>
 
-Vector4::Vector4(double x, double y, double z) {
+Vector4::Vector4(float x, float y, float z) {
   this->data[0] = x;
   this->data[1] = y;
   this->data[2] = z;
   this->data[3] = 0.0;
 }
 
-Vector4::Vector4(double x, double y, double z, double w) {
+Vector4::Vector4(float x, float y, float z, float w) {
   this->data[0] = x;
   this->data[1] = y;
   this->data[2] = z;
@@ -30,16 +30,16 @@ Vector4 Vector4::operator-() const {
 }
 
 //dot product
-double Vector4::operator*(const Vector4 &rhs) const {
-  double x2 = this->data[0] * rhs.data[0];
-  double y2 = this->data[1] * rhs.data[1];
-  double z2 = this->data[2] * rhs.data[2];
-  double w2 = this->data[3] * rhs.data[3];
+float Vector4::operator*(const Vector4 &rhs) const {
+  float x2 = this->data[0] * rhs.data[0];
+  float y2 = this->data[1] * rhs.data[1];
+  float z2 = this->data[2] * rhs.data[2];
+  float w2 = this->data[3] * rhs.data[3];
 
   return (x2 + y2 + z2) ;
 }
 
-Vector4 Vector4::operator*(double t) const {
+Vector4 Vector4::operator*(float t) const {
   return Vector4(t * this->data[0], t * this->data[1], 
                  t * this->data[2], t * this->data[3]);
 }
@@ -51,7 +51,7 @@ Vector4 Vector4::operator|(const Vector4 &v) const {
                  this->data[0] * v.data[1] - this->data[1] * v.data[0]);
 }
 
-Vector4 Vector4::operator/(double t) const {
+Vector4 Vector4::operator/(float t) const {
   return Vector4(this->data[0] / t, 
                  this->data[1] / t,
                  this->data[2] / t,
@@ -59,27 +59,27 @@ Vector4 Vector4::operator/(double t) const {
 }
 
 void Vector4::normalize() {
-    double t = this->length();
+    float t = this->length();
     this->data[0] = this->data[0] / t;
     this->data[1] = this->data[1] / t;
     this->data[2] = this->data[2] / t;
     this->data[3] = this->data[3] / t;
 }
 
-double Vector4::length() const {
-  double x2 = this->data[0] * this->data[0];
-  double y2 = this->data[1] * this->data[1];
-  double z2 = this->data[2] * this->data[2];
-  double w2 = this->data[3] * this->data[3];
+float Vector4::length() const {
+  float x2 = this->data[0] * this->data[0];
+  float y2 = this->data[1] * this->data[1];
+  float z2 = this->data[2] * this->data[2];
+  float w2 = this->data[3] * this->data[3];
   return std::sqrt(x2 + y2 + z2 + w2);
 }
 
 
-Vector4 operator*(double t, const Vector4 &v) {
+Vector4 operator*(float t, const Vector4 &v) {
   return Vector4(t * v.x, t * v.y, t * v.z, t * v.w);
 }
 
-Vector4 operator/(double t, const Vector4 &v) {
+Vector4 operator/(float t, const Vector4 &v) {
   return v / t;
 }
